% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{anime}
\alias{anime}
\title{Anime data set}
\format{
A data frame with 21460 rows and 28 columns

\describe{
\item{id}{abcd}
\item{title}{abcd}
\item{synonyms}{abcd}
\item{synopsis}{abcd}
\item{japanese}{abcd}
\item{english}{type}
}
}
\source{
\url{https://www.kaggle.com/datasets/CooperUnion/anime-recommendations-database}
}
\usage{
anime
}
\description{
Anime data set
}
\keyword{datasets}
