% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_list.R
\name{genre_list}
\alias{genre_list}
\title{The List of Genres}
\usage{
genre_list()
}
\value{
A list of 22 unique genres.
}
\description{
This function helps you to find all genres that you can use as an input for other function, such as \code{anime_rec()}.
}
\examples{

library(animeR)

# Show the list of genres
genre_list()

}
