% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theme.R
\name{get_theme}
\alias{get_theme}
\title{Generate a Word Cloud of the Most Common Themes in the Synopsis}
\usage{
get_theme(input_type)
}
\value{
A word cloud visualization that present the frequency of each the
}
\description{
This function is used to generate a word cloud that shows the most popular words of in all the anime's synopsis.
}
\examples{

library(animeR)

# # Obtain visualizations about current themes in anime movies
get_theme("Movie")

}
