% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend}
\alias{get_trend}
\title{Generate three visualizations regarding duration, score across producers, and score across different genre}
\usage{
get_trend(input_type)
}
\value{
Three visualizations about anime duration, score across different genre and score across different producers.
}
\description{
This function is used to generate visualizations to show the trend of anime duration and score distributions.
}
\examples{

library(animeR)

# Obtain visualizations about current trends in anime movies
get_trend("Movie")

}
