% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_popular.R
\name{rank_popular}
\alias{rank_popular}
\title{animeR}
\usage{
rank_popular(user_num)
}
\arguments{
\item{user_num}{: A numeric vector of the number of anime that user wants as an output.}
}
\value{
A tibble of the specified number of anime and its synopsis, and other information?
}
\description{
This function generates a number (user-specified) of top anime ranked by popularity.
}
\examples{

library(animeR)

# Obtain Top 20 animes ranked by popularity
rank_popular(20)

}
