% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_source_byrating.R
\name{score_source_byrating}
\alias{score_source_byrating}
\title{animeR}
\usage{
score_source_byrating(user_rating)
}
\arguments{
\item{user_rating}{: A character vector of the rating.}
}
\value{
A side-by-side boxplot comparing Score statistics for each Source of animes.
}
\description{
This function produces a side-by-side boxplot between source and score based on a given rating.
}
\examples{

library(animeR)

# Obtain a side-by-side boxplot of Score by Source for animes with rating G.
score_source_byrating("G")


}
