% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_anime.R
\name{search_anime}
\alias{search_anime}
\title{animeR}
\usage{
search_anime(title_short)
}
\arguments{
\item{title_short}{: A character vector of name of the anime title. It can be either full title or a word that is a part of the title.}
}
\value{
A dataframe of title and synopsis of the all the possible anime that user might indicate.
}
\description{
This function can be used as a search engine for anime. It helps you find a specific anime and give its overview, as well as give a list of anime that might be interesting for user and their information.
}
\examples{

library(animeR)

# Obtain the synopsis for all anime that contains "Naruto" in their title
search_anime("Naruto")

}
