% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_list.R
\name{source_list}
\alias{source_list}
\title{The List of Sources}
\usage{
source_list()
}
\value{
A list of 17 unique genres, including Unknown.
}
\description{
This function helps you to find all sources that you can use as an input for other function, such as \code{anime_rec()}.
}
\examples{

library(animeR)

# Show the list of sources
source_list()

}
