% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{stop_words data set}
\format{
A data frame with 21460 rows and 27 columns

\describe{
\item{word}{Common stopwords used in writing}
\item{lexicon}{common stopwords}
}
}
\source{
\url{https://rdrr.io/cran/tidytext/man/stop_words.html}
}
\usage{
stop_words
}
\description{
stop_words data set
}
\keyword{datasets}
