% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_list.R
\name{unique_list}
\alias{unique_list}
\title{The List of Unique Genres, Sources, Themes, Licensors, Producers, or Studios}
\usage{
unique_list(user_variable)
}
\arguments{
\item{user_variable}{: A character vector of a variable's name in the anime dataset. This could be selected from the following list: genres, sources, themes, licensors, producers, and studios.}
}
\value{
A list of unique genres, sources, themes, licensors, producers, or studios.
}
\description{
This function helps you to find all genres and sources that you can use as an input for other function, such as \code{anime_rec()}. In addition to genres and sources, users can check the list of unique themes, licensors, producers, and studios as well.
}
\examples{

library(animeR)

# Show the list of genres
unique_list("genres")

# Show the list of themes
unique_list("themes")

}
