% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_theme.R
\name{view_theme}
\alias{view_theme}
\title{animeR}
\usage{
view_theme()
}
\value{
A word cloud visualization that present the frequency of each the
}
\description{
This function is used to generate a word cloud that shows the most popular words of in all the anime's synopsis.
}
\examples{

library(animeR)

# Obtain a word cloud that shows the most frequent words appeared in all anime's synopsis.
view_theme()

}
