% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{anime}
\alias{anime}
\title{Anime data set}
\format{
\subsection{\code{anime}}{

A data frame with 21460 rows and 28 columns:
\describe{
\item{id}{title}{synonyms}
\item{synopsis}{japanese}
\item{english}{type}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
anime
}
\description{
A list of anime
}
\keyword{datasets}
