% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_rec.R
\name{anime_rec}
\alias{anime_rec}
\title{Find Top 5 Popular Anime}
\usage{
anime_rec(user_genre, user_source)
}
\arguments{
\item{user_genre}{: A character vector of genre of the anime. Users can pick one genre out of the list of 22 genres we provide.}

\item{user_source:}{A character vector of source of the anime. Users can also pick one source out of the list of 17 sources.}
}
\value{
A tibble of top five most popular anime with their titles, synopsis, and popularity.
}
\description{
This function helps you to find the top 5 most popular anime in its certain genre and source. There are 22 unique genres and 17 unique sources to choose from.
}
\examples{

library(animeR)

# Top five most popular action anime that are based on manga
anime_rec("Action", "Manga")

}
