% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend}
\alias{get_trend}
\title{animeR}
\usage{
get_trend()
}
\value{
A word cloud visualization that present the frequency of each the
}
\description{
This function is used to generate a word cloud that shows the most popular words of in all the anime's synopsis.
}
\examples{

library(animeR)

# Obtain a word cloud that shows the most frequent words appeared in all anime's synopsis.
get_trend()

}
