% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_popular.R
\name{rank_popular}
\alias{rank_popular}
\title{animeR}
\usage{
rank_popular()
}
\value{
A dataframe of title and synopsis of the all the possible anime that user might indicate.
}
\description{
This function generates the top 20 anime ranked by popularity.
}
\examples{

library(animeR)

# Obtain Top 20 animes ranked by popularity
rank_popular()

}
