% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_popular.R
\name{rank_popular}
\alias{rank_popular}
\title{animeR}
\usage{
rank_popular(user_num)
}
\arguments{
\item{user_num}{: A numeric vector of the number of anime the user seeks.}
}
\value{
A tibble of the specified number of anime and its genre, source, rank and rating.
}
\description{
This function generates a number (user-specified) of the top animes ranked by popularity.
}
\examples{

library(animeR)

# Obtain Top 20 animes ranked by popularity
rank_popular(20)

}
